refactor: combine namespaces list
