+++ src/components/mixins/settingsGeneralDatabase.ts
@@ -1,17 +1,15 @@
 import Component from 'vue-class-component'
 import BaseMixin from './base'
+import { TranslateResult } from 'vue-i18n'
 
 @Component
 export default class SettingsGeneralDatabase extends BaseMixin {
     get availableKeys() {
         return [
+            // gui namespace
             {
                 value: 'general',
                 label: this.$t('Settings.GeneralTab.General'),
-            },
-            {
-                value: 'console',
-                label: this.$t('Settings.ConsoleTab.Console'),
             },
             {
                 value: 'control',
@@ -28,6 +26,24 @@ export default class SettingsGeneralDatabase extends BaseMixin {
             {
                 value: 'gcodeViewer',
                 label: this.$t('Settings.GCodeViewerTab.GCodeViewer'),
+            },
+            {
+                value: 'navigation',
+                label: this.$t('Settings.GeneralTab.DBNavigation'),
+            },
+            {
+                value: 'uiSettings',
+                label: this.$t('Settings.UiSettingsTab.UiSettings'),
+            },
+            {
+                value: 'view',
+                label: this.$t('Settings.GeneralTab.DbView'),
+            },
+
+            // gui modules
+            {
+                value: 'console',
+                label: this.$t('Settings.ConsoleTab.Console'),
             },
             {
                 value: 'gcodehistory',
@@ -53,14 +69,84 @@ export default class SettingsGeneralDatabase extends BaseMixin {
                 value: 'timelapse',
                 label: this.$t('Settings.TimelapseTab.Timelapse'),
             },
-            {
-                value: 'uiSettings',
-                label: this.$t('Settings.UiSettingsTab.UiSettings'),
-            },
-            {
-                value: 'view',
-                label: this.$t('Settings.GeneralTab.DbView'),
-            },
         ]
+    }
+
+    async loadBackupableNamespaces() {
+        // init namespaces
+        let backupableNamespaces: { value: string; label: string | TranslateResult }[] = []
+
+        // load DB namespaces from moonraker
+        const urlRequestDbList = this.$store.getters['socket/getUrl'] + '/server/database/list'
+        const availableNamespaces = await fetch(urlRequestDbList)
+            // read json
+            .then((response) => response?.json())
+            // extract result namespaces
+            .then((response) => response?.result?.namespaces ?? [])
+            .catch(() => {
+                window.console.error('Cannot load Moonraker DB namespaces')
+                return []
+            })
+
+        // load mainsail keys, if mainsail namespace exists
+        if (availableNamespaces.includes('mainsail')) {
+            const urlRequestMainsailNamespace =
+                this.$store.getters['socket/getUrl'] + '/server/database/item?namespace=mainsail'
+            backupableNamespaces = await fetch(urlRequestMainsailNamespace)
+                // read json
+                .then((response) => response?.json())
+                // extract result object
+                .then((response) => response?.result?.value ?? {})
+                // extract keys from mainsail gui object
+                .then((objects) => Object.keys(objects))
+                // filter initVersion
+                .then((keys) => keys.filter((key) => key !== 'initVersion'))
+                // convert to locale
+                .then((keys) =>
+                    keys.map((key) => {
+                        const namespace = this.availableKeys.find((namespace) => namespace.value === key)
+                        if (namespace) return namespace
+
+                        // fallback return key name
+                        return { value: key, label: key }
+                    })
+                )
+
+            backupableNamespaces = backupableNamespaces.sort(this.sortNamespaces)
+        }
+
+        // add timelapse if exists
+        if (availableNamespaces.includes('timelapse')) {
+            backupableNamespaces.push({
+                value: 'timelapse',
+                label: this.$t('Settings.GeneralTab.DbTimelapseSettings'),
+            })
+        }
+
+        // add webcams if exists
+        if (availableNamespaces.includes('webcams')) {
+            backupableNamespaces.push({
+                value: 'webcams',
+                label: this.$t('Settings.WebcamsTab.Webcams'),
+            })
+        }
+
+        return backupableNamespaces
+    }
+
+    sortNamespaces(
+        a: { value: string; label: string | TranslateResult },
+        b: { value: string; label: string | TranslateResult }
+    ) {
+        if (a.value === 'general') return -1
+        if (b.value === 'general') return 1
+
+        const stringA = a.label.toString().toLowerCase()
+        const stringB = b.label.toString().toLowerCase()
+
+        if (stringA < stringB) return -1
+        if (stringA > stringB) return 1
+
+        return 0
     }
 }
