+++ src/components/settings/General/GeneralReset.vue
@@ -64,79 +64,8 @@ export default class SettingsGeneralTabResetDatabase extends Mixins(BaseMixin, S
     resetableNamespaces: { value: string; label: string | TranslateResult }[] = []
     resetCheckboxes: string[] = []
 
-    mounted() {
-        this.loadResetableNamespaces()
-    }
-
-    async loadResetableNamespaces() {
-        // reset namespaces
-        this.resetableNamespaces = []
-
-        // load DB namespaces from moonraker
-        const urlRequestDbList = this.$store.getters['socket/getUrl'] + '/server/database/list'
-        const availableNamespaces = await fetch(urlRequestDbList)
-            // read json
-            .then((response) => response?.json())
-            // extract result namespaces
-            .then((response) => response?.result?.namespaces ?? [])
-            .catch(() => {
-                window.console.error('Cannot load Moonraker DB namespaces')
-                return []
-            })
-
-        // load mainsail keys, if mainsail namespace exists
-        if (availableNamespaces.includes('mainsail')) {
-            const urlRequestMainsailNamespace =
-                this.$store.getters['socket/getUrl'] + '/server/database/item?namespace=mainsail'
-            this.resetableNamespaces = await fetch(urlRequestMainsailNamespace)
-                // read json
-                .then((response) => response?.json())
-                // extract result object
-                .then((response) => response?.result?.value ?? {})
-                // extract keys from mainsail gui object
-                .then((objects) => Object.keys(objects))
-                // filter initVersion
-                .then((keys) => keys.filter((key) => key !== 'initVersion'))
-                // convert to locale
-                .then((keys) =>
-                    keys.map((key) => {
-                        const namespace = this.availableKeys.find((namespace) => namespace.value === key)
-                        if (namespace) return namespace
-
-                        // fallback return key name
-                        return { value: key, label: key }
-                    })
-                )
-
-            this.resetableNamespaces = this.resetableNamespaces.sort((a, b) => {
-                if (a.value === 'general') return -1
-                if (b.value === 'general') return 1
-
-                const stringA = a.label.toString().toLowerCase()
-                const stringB = b.label.toString().toLowerCase()
-
-                if (stringA < stringB) return -1
-                if (stringA > stringB) return 1
-
-                return 0
-            })
-        }
-
-        // add timelapse if exists
-        if (availableNamespaces.includes('timelapse')) {
-            this.resetableNamespaces.push({
-                value: 'timelapse',
-                label: this.$t('Settings.GeneralTab.DbTimelapseSettings'),
-            })
-        }
-
-        // add webcams if exists
-        if (availableNamespaces.includes('webcams')) {
-            this.resetableNamespaces.push({
-                value: 'webcams',
-                label: this.$t('Settings.GeneralTab.DbWebcams'),
-            })
-        }
+    async mounted() {
+        this.resetableNamespaces = await this.loadBackupableNamespaces()
     }
 
     onSelectResetCheckboxes(resetCheckboxes: string[]) {
@@ -148,8 +77,8 @@ export default class SettingsGeneralTabResetDatabase extends Mixins(BaseMixin, S
         this.$store.dispatch('gui/resetMoonrakerDB', this.resetCheckboxes)
     }
 
-    openDialog() {
-        this.loadResetableNamespaces()
+    async openDialog() {
+        this.resetableNamespaces = await this.loadBackupableNamespaces()
         this.showDialog = true
     }
